deps/v8/test/webkit/legitimately-captured-argument-expected.txt

# Copyright 2013 the V8 project authors. All rights reserved.
# Copyright (C) 2005, 2006, 2007, 2008, 2009 Apple Inc. All rights reserved.
#
# Redistribution and use in source and binary forms, with or without
# modification, are permitted provided that the following conditions
# are met:
# 1.  Redistributions of source code must retain the above copyright
#     notice, this list of conditions and the following disclaimer.
# 2.  Redistributions in binary form must reproduce the above copyright
#     notice, this list of conditions and the following disclaimer in the
#     documentation and/or other materials provided with the distribution.
#
# THIS SOFTWARE IS PROVIDED BY APPLE INC. AND ITS CONTRIBUTORS ``AS IS'' AND ANY
# EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
# WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
# DISCLAIMED. IN NO EVENT SHALL APPLE INC. OR ITS CONTRIBUTORS BE LIABLE FOR ANY
# DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
# (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
# LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
# ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
# (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
# SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Tests that using an argument as a captured variable, in the legitimate sense rather than the function.arguments sense, works as expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS counter() is 1
PASS counter() is 2
PASS counter() is 3
PASS counter() is 4
PASS counter() is 5
PASS counter() is 6
PASS counter() is 7
PASS counter() is 8
PASS counter() is 9
PASS counter() is 10
PASS counter() is 2
PASS counter() is 3
PASS counter() is 4
PASS counter() is 5
PASS counter() is 6
PASS counter() is 7
PASS counter() is 8
PASS counter() is 9
PASS counter() is 10
PASS counter() is 11
PASS counter() is 3
PASS counter() is 4
PASS counter() is 5
PASS counter() is 6
PASS counter() is 7
PASS counter() is 8
PASS counter() is 9
PASS counter() is 10
PASS counter() is 11
PASS counter() is 12
PASS counter() is 4
PASS counter() is 5
PASS counter() is 6
PASS counter() is 7
PASS counter() is 8
PASS counter() is 9
PASS counter() is 10
PASS counter() is 11
PASS counter() is 12
PASS counter() is 13
PASS counter() is 5
PASS counter() is 6
PASS counter() is 7
PASS counter() is 8
PASS counter() is 9
PASS counter() is 10
PASS counter() is 11
PASS counter() is 12
PASS counter() is 13
PASS counter() is 14
PASS counter() is 6
PASS counter() is 7
PASS counter() is 8
PASS counter() is 9
PASS counter() is 10
PASS counter() is 11
PASS counter() is 12
PASS counter() is 13
PASS counter() is 14
PASS counter() is 15
PASS counter() is 7
PASS counter() is 8
PASS counter() is 9
PASS counter() is 10
PASS counter() is 11
PASS counter() is 12
PASS counter() is 13
PASS counter() is 14
PASS counter() is 15
PASS counter() is 16
PASS counter() is 8
PASS counter() is 9
PASS counter() is 10
PASS counter() is 11
PASS counter() is 12
PASS counter() is 13
PASS counter() is 14
PASS counter() is 15
PASS counter() is 16
PASS counter() is 17
PASS counter() is 9
PASS counter() is 10
PASS counter() is 11
PASS counter() is 12
PASS counter() is 13
PASS counter() is 14
PASS counter() is 15
PASS counter() is 16
PASS counter() is 17
PASS counter() is 18
PASS counter() is 10
PASS counter() is 11
PASS counter() is 12
PASS counter() is 13
PASS counter() is 14
PASS counter() is 15
PASS counter() is 16
PASS counter() is 17
PASS counter() is 18
PASS counter() is 19
PASS counter() is 11
PASS counter() is 12
PASS counter() is 13
PASS counter() is 14
PASS counter() is 15
PASS counter() is 16
PASS counter() is 17
PASS counter() is 18
PASS counter() is 19
PASS counter() is 20
PASS counter() is 12
PASS counter() is 13
PASS counter() is 14
PASS counter() is 15
PASS counter() is 16
PASS counter() is 17
PASS counter() is 18
PASS counter() is 19
PASS counter() is 20
PASS counter() is 21
PASS counter() is 13
PASS counter() is 14
PASS counter() is 15
PASS counter() is 16
PASS counter() is 17
PASS counter() is 18
PASS counter() is 19
PASS counter() is 20
PASS counter() is 21
PASS counter() is 22
PASS counter() is 14
PASS counter() is 15
PASS counter() is 16
PASS counter() is 17
PASS counter() is 18
PASS counter() is 19
PASS counter() is 20
PASS counter() is 21
PASS counter() is 22
PASS counter() is 23
PASS counter() is 15
PASS counter() is 16
PASS counter() is 17
PASS counter() is 18
PASS counter() is 19
PASS counter() is 20
PASS counter() is 21
PASS counter() is 22
PASS counter() is 23
PASS counter() is 24
PASS counter() is 16
PASS counter() is 17
PASS counter() is 18
PASS counter() is 19
PASS counter() is 20
PASS counter() is 21
PASS counter() is 22
PASS counter() is 23
PASS counter() is 24
PASS counter() is 25
PASS counter() is 17
PASS counter() is 18
PASS counter() is 19
PASS counter() is 20
PASS counter() is 21
PASS counter() is 22
PASS counter() is 23
PASS counter() is 24
PASS counter() is 25
PASS counter() is 26
PASS counter() is 18
PASS counter() is 19
PASS counter() is 20
PASS counter() is 21
PASS counter() is 22
PASS counter() is 23
PASS counter() is 24
PASS counter() is 25
PASS counter() is 26
PASS counter() is 27
PASS counter() is 19
PASS counter() is 20
PASS counter() is 21
PASS counter() is 22
PASS counter() is 23
PASS counter() is 24
PASS counter() is 25
PASS counter() is 26
PASS counter() is 27
PASS counter() is 28
PASS counter() is 20
PASS counter() is 21
PASS counter() is 22
PASS counter() is 23
PASS counter() is 24
PASS counter() is 25
PASS counter() is 26
PASS counter() is 27
PASS counter() is 28
PASS counter() is 29
PASS counter() is 21
PASS counter() is 22
PASS counter() is 23
PASS counter() is 24
PASS counter() is 25
PASS counter() is 26
PASS counter() is 27
PASS counter() is 28
PASS counter() is 29
PASS counter() is 30
PASS counter() is 22
PASS counter() is 23
PASS counter() is 24
PASS counter() is 25
PASS counter() is 26
PASS counter() is 27
PASS counter() is 28
PASS counter() is 29
PASS counter() is 30
PASS counter() is 31
PASS counter() is 23
PASS counter() is 24
PASS counter() is 25
PASS counter() is 26
PASS counter() is 27
PASS counter() is 28
PASS counter() is 29
PASS counter() is 30
PASS counter() is 31
PASS counter() is 32
PASS counter() is 24
PASS counter() is 25
PASS counter() is 26
PASS counter() is 27
PASS counter() is 28
PASS counter() is 29
PASS counter() is 30
PASS counter() is 31
PASS counter() is 32
PASS counter() is 33
PASS counter() is 25
PASS counter() is 26
PASS counter() is 27
PASS counter() is 28
PASS counter() is 29
PASS counter() is 30
PASS counter() is 31
PASS counter() is 32
PASS counter() is 33
PASS counter() is 34
PASS counter() is 26
PASS counter() is 27
PASS counter() is 28
PASS counter() is 29
PASS counter() is 30
PASS counter() is 31
PASS counter() is 32
PASS counter() is 33
PASS counter() is 34
PASS counter() is 35
PASS counter() is 27
PASS counter() is 28
PASS counter() is 29
PASS counter() is 30
PASS counter() is 31
PASS counter() is 32
PASS counter() is 33
PASS counter() is 34
PASS counter() is 35
PASS counter() is 36
PASS counter() is 28
PASS counter() is 29
PASS counter() is 30
PASS counter() is 31
PASS counter() is 32
PASS counter() is 33
PASS counter() is 34
PASS counter() is 35
PASS counter() is 36
PASS counter() is 37
PASS counter() is 29
PASS counter() is 30
PASS counter() is 31
PASS counter() is 32
PASS counter() is 33
PASS counter() is 34
PASS counter() is 35
PASS counter() is 36
PASS counter() is 37
PASS counter() is 38
PASS counter() is 30
PASS counter() is 31
PASS counter() is 32
PASS counter() is 33
PASS counter() is 34
PASS counter() is 35
PASS counter() is 36
PASS counter() is 37
PASS counter() is 38
PASS counter() is 39
PASS counter() is 31
PASS counter() is 32
PASS counter() is 33
PASS counter() is 34
PASS counter() is 35
PASS counter() is 36
PASS counter() is 37
PASS counter() is 38
PASS counter() is 39
PASS counter() is 40
PASS counter() is 32
PASS counter() is 33
PASS counter() is 34
PASS counter() is 35
PASS counter() is 36
PASS counter() is 37
PASS counter() is 38
PASS counter() is 39
PASS counter() is 40
PASS counter() is 41
PASS counter() is 33
PASS counter() is 34
PASS counter() is 35
PASS counter() is 36
PASS counter() is 37
PASS counter() is 38
PASS counter() is 39
PASS counter() is 40
PASS counter() is 41
PASS counter() is 42
PASS counter() is 34
PASS counter() is 35
PASS counter() is 36
PASS counter() is 37
PASS counter() is 38
PASS counter() is 39
PASS counter() is 40
PASS counter() is 41
PASS counter() is 42
PASS counter() is 43
PASS counter() is 35
PASS counter() is 36
PASS counter() is 37
PASS counter() is 38
PASS counter() is 39
PASS counter() is 40
PASS counter() is 41
PASS counter() is 42
PASS counter() is 43
PASS counter() is 44
PASS counter() is 36
PASS counter() is 37
PASS counter() is 38
PASS counter() is 39
PASS counter() is 40
PASS counter() is 41
PASS counter() is 42
PASS counter() is 43
PASS counter() is 44
PASS counter() is 45
PASS counter() is 37
PASS counter() is 38
PASS counter() is 39
PASS counter() is 40
PASS counter() is 41
PASS counter() is 42
PASS counter() is 43
PASS counter() is 44
PASS counter() is 45
PASS counter() is 46
PASS counter() is 38
PASS counter() is 39
PASS counter() is 40
PASS counter() is 41
PASS counter() is 42
PASS counter() is 43
PASS counter() is 44
PASS counter() is 45
PASS counter() is 46
PASS counter() is 47
PASS counter() is 39
PASS counter() is 40
PASS counter() is 41
PASS counter() is 42
PASS counter() is 43
PASS counter() is 44
PASS counter() is 45
PASS counter() is 46
PASS counter() is 47
PASS counter() is 48
PASS counter() is 40
PASS counter() is 41
PASS counter() is 42
PASS counter() is 43
PASS counter() is 44
PASS counter() is 45
PASS counter() is 46
PASS counter() is 47
PASS counter() is 48
PASS counter() is 49
PASS counter() is 41
PASS counter() is 42
PASS counter() is 43
PASS counter() is 44
PASS counter() is 45
PASS counter() is 46
PASS counter() is 47
PASS counter() is 48
PASS counter() is 49
PASS counter() is 50
PASS counter() is 42
PASS counter() is 43
PASS counter() is 44
PASS counter() is 45
PASS counter() is 46
PASS counter() is 47
PASS counter() is 48
PASS counter() is 49
PASS counter() is 50
PASS counter() is 51
PASS counter() is 43
PASS counter() is 44
PASS counter() is 45
PASS counter() is 46
PASS counter() is 47
PASS counter() is 48
PASS counter() is 49
PASS counter() is 50
PASS counter() is 51
PASS counter() is 52
PASS counter() is 44
PASS counter() is 45
PASS counter() is 46
PASS counter() is 47
PASS counter() is 48
PASS counter() is 49
PASS counter() is 50
PASS counter() is 51
PASS counter() is 52
PASS counter() is 53
PASS counter() is 45
PASS counter() is 46
PASS counter() is 47
PASS counter() is 48
PASS counter() is 49
PASS counter() is 50
PASS counter() is 51
PASS counter() is 52
PASS counter() is 53
PASS counter() is 54
PASS counter() is 46
PASS counter() is 47
PASS counter() is 48
PASS counter() is 49
PASS counter() is 50
PASS counter() is 51
PASS counter() is 52
PASS counter() is 53
PASS counter() is 54
PASS counter() is 55
PASS counter() is 47
PASS counter() is 48
PASS counter() is 49
PASS counter() is 50
PASS counter() is 51
PASS counter() is 52
PASS counter() is 53
PASS counter() is 54
PASS counter() is 55
PASS counter() is 56
PASS counter() is 48
PASS counter() is 49
PASS counter() is 50
PASS counter() is 51
PASS counter() is 52
PASS counter() is 53
PASS counter() is 54
PASS counter() is 55
PASS counter() is 56
PASS counter() is 57
PASS counter() is 49
PASS counter() is 50
PASS counter() is 51
PASS counter() is 52
PASS counter() is 53
PASS counter() is 54
PASS counter() is 55
PASS counter() is 56
PASS counter() is 57
PASS counter() is 58
PASS counter() is 50
PASS counter() is 51
PASS counter() is 52
PASS counter() is 53
PASS counter() is 54
PASS counter() is 55
PASS counter() is 56
PASS counter() is 57
PASS counter() is 58
PASS counter() is 59
PASS counter() is 51
PASS counter() is 52
PASS counter() is 53
PASS counter() is 54
PASS counter() is 55
PASS counter() is 56
PASS counter() is 57
PASS counter() is 58
PASS counter() is 59
PASS counter() is 60
PASS counter() is 52
PASS counter() is 53
PASS counter() is 54
PASS counter() is 55
PASS counter() is 56
PASS counter() is 57
PASS counter() is 58
PASS counter() is 59
PASS counter() is 60
PASS counter() is 61
PASS counter() is 53
PASS counter() is 54
PASS counter() is 55
PASS counter() is 56
PASS counter() is 57
PASS counter() is 58
PASS counter() is 59
PASS counter() is 60
PASS counter() is 61
PASS counter() is 62
PASS counter() is 54
PASS counter() is 55
PASS counter() is 56
PASS counter() is 57
PASS counter() is 58
PASS counter() is 59
PASS counter() is 60
PASS counter() is 61
PASS counter() is 62
PASS counter() is 63
PASS counter() is 55
PASS counter() is 56
PASS counter() is 57
PASS counter() is 58
PASS counter() is 59
PASS counter() is 60
PASS counter() is 61
PASS counter() is 62
PASS counter() is 63
PASS counter() is 64
PASS counter() is 56
PASS counter() is 57
PASS counter() is 58
PASS counter() is 59
PASS counter() is 60
PASS counter() is 61
PASS counter() is 62
PASS counter() is 63
PASS counter() is 64
PASS counter() is 65
PASS counter() is 57
PASS counter() is 58
PASS counter() is 59
PASS counter() is 60
PASS counter() is 61
PASS counter() is 62
PASS counter() is 63
PASS counter() is 64
PASS counter() is 65
PASS counter() is 66
PASS counter() is 58
PASS counter() is 59
PASS counter() is 60
PASS counter() is 61
PASS counter() is 62
PASS counter() is 63
PASS counter() is 64
PASS counter() is 65
PASS counter() is 66
PASS counter() is 67
PASS counter() is 59
PASS counter() is 60
PASS counter() is 61
PASS counter() is 62
PASS counter() is 63
PASS counter() is 64
PASS counter() is 65
PASS counter() is 66
PASS counter() is 67
PASS counter() is 68
PASS counter() is 60
PASS counter() is 61
PASS counter() is 62
PASS counter() is 63
PASS counter() is 64
PASS counter() is 65
PASS counter() is 66
PASS counter() is 67
PASS counter() is 68
PASS counter() is 69
PASS counter() is 61
PASS counter() is 62
PASS counter() is 63
PASS counter() is 64
PASS counter() is 65
PASS counter() is 66
PASS counter() is 67
PASS counter() is 68
PASS counter() is 69
PASS counter() is 70
PASS counter() is 62
PASS counter() is 63
PASS counter() is 64
PASS counter() is 65
PASS counter() is 66
PASS counter() is 67
PASS counter() is 68
PASS counter() is 69
PASS counter() is 70
PASS counter() is 71
PASS counter() is 63
PASS counter() is 64
PASS counter() is 65
PASS counter() is 66
PASS counter() is 67
PASS counter() is 68
PASS counter() is 69
PASS counter() is 70
PASS counter() is 71
PASS counter() is 72
PASS counter() is 64
PASS counter() is 65
PASS counter() is 66
PASS counter() is 67
PASS counter() is 68
PASS counter() is 69
PASS counter() is 70
PASS counter() is 71
PASS counter() is 72
PASS counter() is 73
PASS counter() is 65
PASS counter() is 66
PASS counter() is 67
PASS counter() is 68
PASS counter() is 69
PASS counter() is 70
PASS counter() is 71
PASS counter() is 72
PASS counter() is 73
PASS counter() is 74
PASS counter() is 66
PASS counter() is 67
PASS counter() is 68
PASS counter() is 69
PASS counter() is 70
PASS counter() is 71
PASS counter() is 72
PASS counter() is 73
PASS counter() is 74
PASS counter() is 75
PASS counter() is 67
PASS counter() is 68
PASS counter() is 69
PASS counter() is 70
PASS counter() is 71
PASS counter() is 72
PASS counter() is 73
PASS counter() is 74
PASS counter() is 75
PASS counter() is 76
PASS counter() is 68
PASS counter() is 69
PASS counter() is 70
PASS counter() is 71
PASS counter() is 72
PASS counter() is 73
PASS counter() is 74
PASS counter() is 75
PASS counter() is 76
PASS counter() is 77
PASS counter() is 69
PASS counter() is 70
PASS counter() is 71
PASS counter() is 72
PASS counter() is 73
PASS counter() is 74
PASS counter() is 75
PASS counter() is 76
PASS counter() is 77
PASS counter() is 78
PASS counter() is 70
PASS counter() is 71
PASS counter() is 72
PASS counter() is 73
PASS counter() is 74
PASS counter() is 75
PASS counter() is 76
PASS counter() is 77
PASS counter() is 78
PASS counter() is 79
PASS counter() is 71
PASS counter() is 72
PASS counter() is 73
PASS counter() is 74
PASS counter() is 75
PASS counter() is 76
PASS counter() is 77
PASS counter() is 78
PASS counter() is 79
PASS counter() is 80
PASS counter() is 72
PASS counter() is 73
PASS counter() is 74
PASS counter() is 75
PASS counter() is 76
PASS counter() is 77
PASS counter() is 78
PASS counter() is 79
PASS counter() is 80
PASS counter() is 81
PASS counter() is 73
PASS counter() is 74
PASS counter() is 75
PASS counter() is 76
PASS counter() is 77
PASS counter() is 78
PASS counter() is 79
PASS counter() is 80
PASS counter() is 81
PASS counter() is 82
PASS counter() is 74
PASS counter() is 75
PASS counter() is 76
PASS counter() is 77
PASS counter() is 78
PASS counter() is 79
PASS counter() is 80
PASS counter() is 81
PASS counter() is 82
PASS counter() is 83
PASS counter() is 75
PASS counter() is 76
PASS counter() is 77
PASS counter() is 78
PASS counter() is 79
PASS counter() is 80
PASS counter() is 81
PASS counter() is 82
PASS counter() is 83
PASS counter() is 84
PASS counter() is 76
PASS counter() is 77
PASS counter() is 78
PASS counter() is 79
PASS counter() is 80
PASS counter() is 81
PASS counter() is 82
PASS counter() is 83
PASS counter() is 84
PASS counter() is 85
PASS counter() is 77
PASS counter() is 78
PASS counter() is 79
PASS counter() is 80
PASS counter() is 81
PASS counter() is 82
PASS counter() is 83
PASS counter() is 84
PASS counter() is 85
PASS counter() is 86
PASS counter() is 78
PASS counter() is 79
PASS counter() is 80
PASS counter() is 81
PASS counter() is 82
PASS counter() is 83
PASS counter() is 84
PASS counter() is 85
PASS counter() is 86
PASS counter() is 87
PASS counter() is 79
PASS counter() is 80
PASS counter() is 81
PASS counter() is 82
PASS counter() is 83
PASS counter() is 84
PASS counter() is 85
PASS counter() is 86
PASS counter() is 87
PASS counter() is 88
PASS counter() is 80
PASS counter() is 81
PASS counter() is 82
PASS counter() is 83
PASS counter() is 84
PASS counter() is 85
PASS counter() is 86
PASS counter() is 87
PASS counter() is 88
PASS counter() is 89
PASS counter() is 81
PASS counter() is 82
PASS counter() is 83
PASS counter() is 84
PASS counter() is 85
PASS counter() is 86
PASS counter() is 87
PASS counter() is 88
PASS counter() is 89
PASS counter() is 90
PASS counter() is 82
PASS counter() is 83
PASS counter() is 84
PASS counter() is 85
PASS counter() is 86
PASS counter() is 87
PASS counter() is 88
PASS counter() is 89
PASS counter() is 90
PASS counter() is 91
PASS counter() is 83
PASS counter() is 84
PASS counter() is 85
PASS counter() is 86
PASS counter() is 87
PASS counter() is 88
PASS counter() is 89
PASS counter() is 90
PASS counter() is 91
PASS counter() is 92
PASS counter() is 84
PASS counter() is 85
PASS counter() is 86
PASS counter() is 87
PASS counter() is 88
PASS counter() is 89
PASS counter() is 90
PASS counter() is 91
PASS counter() is 92
PASS counter() is 93
PASS counter() is 85
PASS counter() is 86
PASS counter() is 87
PASS counter() is 88
PASS counter() is 89
PASS counter() is 90
PASS counter() is 91
PASS counter() is 92
PASS counter() is 93
PASS counter() is 94
PASS counter() is 86
PASS counter() is 87
PASS counter() is 88
PASS counter() is 89
PASS counter() is 90
PASS counter() is 91
PASS counter() is 92
PASS counter() is 93
PASS counter() is 94
PASS counter() is 95
PASS counter() is 87
PASS counter() is 88
PASS counter() is 89
PASS counter() is 90
PASS counter() is 91
PASS counter() is 92
PASS counter() is 93
PASS counter() is 94
PASS counter() is 95
PASS counter() is 96
PASS counter() is 88
PASS counter() is 89
PASS counter() is 90
PASS counter() is 91
PASS counter() is 92
PASS counter() is 93
PASS counter() is 94
PASS counter() is 95
PASS counter() is 96
PASS counter() is 97
PASS counter() is 89
PASS counter() is 90
PASS counter() is 91
PASS counter() is 92
PASS counter() is 93
PASS counter() is 94
PASS counter() is 95
PASS counter() is 96
PASS counter() is 97
PASS counter() is 98
PASS counter() is 90
PASS counter() is 91
PASS counter() is 92
PASS counter() is 93
PASS counter() is 94
PASS counter() is 95
PASS counter() is 96
PASS counter() is 97
PASS counter() is 98
PASS counter() is 99
PASS counter() is 91
PASS counter() is 92
PASS counter() is 93
PASS counter() is 94
PASS counter() is 95
PASS counter() is 96
PASS counter() is 97
PASS counter() is 98
PASS counter() is 99
PASS counter() is 100
PASS counter() is 92
PASS counter() is 93
PASS counter() is 94
PASS counter() is 95
PASS counter() is 96
PASS counter() is 97
PASS counter() is 98
PASS counter() is 99
PASS counter() is 100
PASS counter() is 101
PASS counter() is 93
PASS counter() is 94
PASS counter() is 95
PASS counter() is 96
PASS counter() is 97
PASS counter() is 98
PASS counter() is 99
PASS counter() is 100
PASS counter() is 101
PASS counter() is 102
PASS counter() is 94
PASS counter() is 95
PASS counter() is 96
PASS counter() is 97
PASS counter() is 98
PASS counter() is 99
PASS counter() is 100
PASS counter() is 101
PASS counter() is 102
PASS counter() is 103
PASS counter() is 95
PASS counter() is 96
PASS counter() is 97
PASS counter() is 98
PASS counter() is 99
PASS counter() is 100
PASS counter() is 101
PASS counter() is 102
PASS counter() is 103
PASS counter() is 104
PASS counter() is 96
PASS counter() is 97
PASS counter() is 98
PASS counter() is 99
PASS counter() is 100
PASS counter() is 101
PASS counter() is 102
PASS counter() is 103
PASS counter() is 104
PASS counter() is 105
PASS counter() is 97
PASS counter() is 98
PASS counter() is 99
PASS counter() is 100
PASS counter() is 101
PASS counter() is 102
PASS counter() is 103
PASS counter() is 104
PASS counter() is 105
PASS counter() is 106
PASS counter() is 98
PASS counter() is 99
PASS counter() is 100
PASS counter() is 101
PASS counter() is 102
PASS counter() is 103
PASS counter() is 104
PASS counter() is 105
PASS counter() is 106
PASS counter() is 107
PASS counter() is 99
PASS counter() is 100
PASS counter() is 101
PASS counter() is 102
PASS counter() is 103
PASS counter() is 104
PASS counter() is 105
PASS counter() is 106
PASS counter() is 107
PASS counter() is 108
PASS counter() is 100
PASS counter() is 101
PASS counter() is 102
PASS counter() is 103
PASS counter() is 104
PASS counter() is 105
PASS counter() is 106
PASS counter() is 107
PASS counter() is 108
PASS counter() is 109
PASS successfullyParsed is true

TEST COMPLETE
